refactor(web): replace SoftwareSection with SoftwareSummary
