+++ web/src/components/overview/SoftwareSection.jsx
@@ -19,135 +19,58 @@
  * find current contact information at www.suse.com.
  */
 
-import React, { useEffect, useReducer } from "react";
-import { BUSY } from "~/client/status";
-import { Button } from "@patternfly/react-core";
-import { Icon } from "~/components/layout";
-import { ProgressText, Section } from "~/components/core";
-import { toValidationError, useCancellablePromise } from "~/utils";
-import { UsedSize } from "~/components/software";
-import { useInstallerClient } from "~/context/installer";
+import React, { useEffect, useState } from "react";
 import { _ } from "~/i18n";
+import { useInstallerClient } from "~/context/installer";
+import { List, ListItem, Text, TextContent, TextVariants } from "@patternfly/react-core";
+import { Em } from "~/components/core";
 
-const initialState = {
-  busy: true,
-  errors: [],
-  errorsRead: false,
-  size: "",
-  progress: { message: _("Reading software repositories"), current: 0, total: 0, finished: false },
-};
+export default function SoftwareSection() {
+  const [proposal, setProposal] = useState({});
+  const [patterns, setPatterns] = useState([]);
+  const [selectedPatterns, setSelectedPatterns] = useState(undefined);
+  const client = useInstallerClient();
 
-const reducer = (state, action) => {
-  switch (action.type) {
-    case "UPDATE_PROGRESS": {
-      const { message, current, total, finished } = action.payload;
-      return { ...state, progress: { message, current, total, finished } };
-    }
+  useEffect(() => {
+    client.software.getProposal().then(setProposal);
+    client.software.getPatterns().then(setPatterns);
+  }, [client]);
 
-    case "UPDATE_STATUS": {
-      return { ...initialState, busy: action.payload.status === BUSY };
-    }
+  useEffect(() => {
+    return client.software.onSelectedPatternsChanged(() => {
+      client.software.getProposal().then(setProposal);
+    });
+  }, [client, setProposal]);
 
-    case "UPDATE_PROPOSAL": {
-      if (state.busy) return state;
+  useEffect(() => {
+    if (proposal.patterns === undefined) return;
 
-      const { errors, size } = action.payload;
-      return { ...state, errors, size, errorsRead: true };
-    }
+    const ids = Object.keys(proposal.patterns);
+    const selected = patterns.filter(p => ids.includes(p.name)).sort((a, b) => a.order - b.order);
+    setSelectedPatterns(selected);
+  }, [client, proposal, patterns]);
 
-    default: {
-      return state;
-    }
+  if (selectedPatterns === undefined) {
+    return;
   }
-};
 
-export default function SoftwareSection({ showErrors }) {
-  const { software: client } = useInstallerClient();
-  const { cancellablePromise } = useCancellablePromise();
-  const [state, dispatch] = useReducer(reducer, initialState);
-
-  const updateStatus = (status) => {
-    dispatch({ type: "UPDATE_STATUS", payload: { status } });
-  };
-
-  const probe = () => client.probe();
-
-  useEffect(() => {
-    cancellablePromise(client.getStatus()).then(updateStatus);
-
-    return client.onStatusChange(updateStatus);
-  }, [client, cancellablePromise]);
-
-  useEffect(() => {
-    if (state.busy) return;
-
-    const updateProposal = async () => {
-      const errors = await cancellablePromise(client.getIssues());
-      const { size } = await cancellablePromise(client.getProposal());
-
-      dispatch({ type: "UPDATE_PROPOSAL", payload: { errors, size } });
-    };
-
-    updateProposal();
-  }, [client, cancellablePromise, state.busy]);
-
-  useEffect(() => {
-    cancellablePromise(client.getProgress()).then(({ message, current, total, finished }) => {
-      dispatch({
-        type: "UPDATE_PROGRESS",
-        payload: { message, current, total, finished },
-      });
-    });
-  }, [client, cancellablePromise]);
-
-  useEffect(() => {
-    return client.onProgressChange(({ message, current, total, finished }) => {
-      dispatch({
-        type: "UPDATE_PROGRESS",
-        payload: { message, current, total, finished },
-      });
-    });
-  }, [client, cancellablePromise]);
-
-  const errors = showErrors ? state.errors : [];
-
-  const SectionContent = () => {
-    if (state.busy) {
-      const { message, current, total } = state.progress;
-      return <ProgressText message={message} current={current} total={total} />;
-    }
-
-    return (
-      <>
-        <UsedSize size={state.size} />
-        {errors.length > 0 &&
-          (
-            <Button
-              isInline
-              variant="link"
-              icon={<Icon name="refresh" size="xxs" />}
-              onClick={probe}
-            >
-              {/* TRANSLATORS: clickable link label */}
-              {_("Refresh the repositories")}
-            </Button>
-          )}
-      </>
-    );
-  };
+  // TRANSLATORS: %s will be replaced with the installation size, example:
+  // "5GiB".
+  const [msg1, msg2] = _("The installation will take %s including:").split("%s");
 
   return (
-    <Section
-      key="software-section"
-      // TRANSLATORS: page section
-      title={_("Software")}
-      icon="apps"
-      loading={state.busy}
-      errors={errors.map(toValidationError)}
-      path="/software"
-      id="software"
-    >
-      <SectionContent />
-    </Section>
+    <TextContent>
+      <Text component={TextVariants.h3}>{_("Software")}</Text>
+      <Text>
+        {msg1}
+        <Em>{`${proposal.size}`}</Em>
+        {msg2}
+      </Text>
+      <List>
+        {selectedPatterns.map(p => (
+          <ListItem key={p.name}>{p.description}</ListItem>
+        ))}
+      </List>
+    </TextContent>
   );
 }
